Example Data for the CSCI tool

A list of two data frames, corresponding to the "bugs" type data frame (benthic macroinvertebrate count data) and "stations" type data frame (GIS-based site data) for the CSCI function. This is a toy data set and should not be used to make any inferences about the health of the included sites.

The bugs data frame contains the following:

StationCode: A code for the location of the sample

SampleID: A unique ID for the sample

FinalID: Names for the taxa identified in each sample. The names must correspond to a SWAMP FinalID (see http://swamp.mpsl.mlml.calstate.edu/).

BAResult: A count for that FinalID in that SampleID. Must be a non-negative integer.

LifeStageCode: Code for the life stage of the taxon. May be "A", "L", or "P" for insects; all non-insects must be "X".

Distinct: Taxonomist's distinctiveness designation. This will override the automatic distinctiveness designations made by the CSCI function. 1 indicates distinctiveness, 0 or NA will defer to the function's designation.

The stations data frame contains the following:

StationCode: A code for the location of the sample. Every StationCode in the bugs data frame is also represented here.

AREA_SQKM: The area of the watershed in square kilometers

New_Long: The site's longitude

New_Lat: The site's latitude

SITE_ELEV: Elevation in ???

TEMP_00_09: The average temperature from 2000 to 2009 in hundredths of degrees C

PPT_00_09: The average precipatation from 2000 to 2009 in hundredths of millimeters

SumAve_P: ???

KFCT_AVE: ???

BDH_AVE: ???

MgO_Mean: Average soil magnesium oxide in ???

P_MEAN: ???

CaO_Mean: Average calcareous soils in ???

PRMH_AVE: ???

S_Mean: ???

PCT_SEDIM: Percent sedimentary geology

LRPREM_mean: ???

N_MEAN: ???

LogWSA: log10 of the AREA_SQKM. Not needed if AREA_SQKM is present.

See also